chore: split tests for readability/cleanup
